make a unified RTC driver for all chip types with both timer & clock capabilities
