+++ boards/feather_m4/examples/sleeping_timer_rtc.rs
@@ -1,26 +1,15 @@
 //! Uses an RTC in standby mode to blink an LED for maximum power savings.
-//!
-//! We also use the internal 8MHz RC oscillator & the internal 32k oscillator
-//! to save power as it doesn't require the PLL. In standby on my feather_m0,
-//! we draw 318 uA when the LED is off and 965 uA when the LED is on at 4.3v.
-//! We are slightly limited by the quiescent current draw of the regulator of
-//! 90-150 uA.
-//!
-//! Some more power numbers that I measured:
-//! In IDLE2 we draw 740 uA (CPU & AHB & APB domain off)
-//! In IDLE1 we draw 870 uA (CPU & AHB domain off)
-//! In IDLE0 we draw 870 uA (CPU domain off)
 #![no_std]
 #![no_main]
 
 extern crate cortex_m;
-extern crate feather_m0 as hal;
+extern crate feather_m4 as hal;
 #[cfg(not(feature = "use_semihosting"))]
 extern crate panic_halt;
 #[cfg(feature = "use_semihosting")]
 extern crate panic_semihosting;
 
-use hal::clock::{enable_internal_32kosc, ClockGenId, ClockSource, GenericClockController};
+use hal::clock::GenericClockController;
 use hal::entry;
 use hal::pac::{interrupt, CorePeripherals, Peripherals, RTC};
 use hal::prelude::*;
@@ -38,22 +27,16 @@ fn main() -> ! {
     // Configure all of our peripherals/clocks
     let mut peripherals = Peripherals::take().unwrap();
     let mut core = CorePeripherals::take().unwrap();
-    let mut clocks = GenericClockController::with_internal_8mhz(
+    let _clocks = GenericClockController::with_internal_32kosc(
         peripherals.GCLK,
-        &mut peripherals.PM,
-        &mut peripherals.SYSCTRL,
+        &mut peripherals.MCLK,
+        &mut peripherals.OSC32KCTRL,
+        &mut peripherals.OSCCTRL,
         &mut peripherals.NVMCTRL,
     );
 
-    // Get a clock & make a sleeping delay object. use internal 32k clock that runs
-    // in standby
-    enable_internal_32kosc(&mut peripherals.SYSCTRL);
-    let timer_clock = clocks
-        .configure_gclk_divider_and_source(ClockGenId::GCLK1, 1, ClockSource::OSC32K, false)
-        .unwrap();
-    clocks.configure_standby(ClockGenId::GCLK1, true);
-    let rtc_clock = clocks.rtc(&timer_clock).unwrap();
-    let timer = rtc::Rtc::new(peripherals.RTC, rtc_clock.freq(), &mut peripherals.PM);
+    // Configure the RTC. a 1024 Hz clock is configured for us when enabling our main clock
+    let timer = rtc::Rtc::new(peripherals.RTC, 1024.hz(), &mut peripherals.MCLK);
     let mut sleeping_delay = SleepingDelay::new(timer, &INTERRUPT_FIRED);
 
     // We can use the RTC in standby for maximum power savings
@@ -66,25 +49,19 @@ fn main() -> ! {
     }
 
     // Turn off unnecessary peripherals
-    peripherals.PM.ahbmask.modify(|_, w| {
+    peripherals.MCLK.ahbmask.modify(|_, w| {
         w.usb_().clear_bit();
         w.dmac_().clear_bit()
     });
-    peripherals.PM.apbamask.modify(|_, w| {
+    peripherals.MCLK.apbamask.modify(|_, w| {
         w.eic_().clear_bit();
-        w.wdt_().clear_bit();
-        w.sysctrl_().clear_bit();
-        w.pac0_().clear_bit()
+        w.wdt_().clear_bit()
     });
-    peripherals.PM.apbbmask.modify(|_, w| {
+    peripherals.MCLK.apbbmask.modify(|_, w| {
         w.usb_().clear_bit();
-        w.dmac_().clear_bit();
         w.nvmctrl_().clear_bit();
-        w.dsu_().clear_bit();
-        w.pac1_().clear_bit()
+        w.dsu_().clear_bit()
     });
-    // Thankfully the only one default on here is ADC
-    peripherals.PM.apbcmask.modify(|_, w| w.adc_().clear_bit());
 
     // Configure our red LED and blink forever, sleeping between!
     let mut pins = hal::Pins::new(peripherals.PORT);
