+++ hal/src/common/thumbv6m/mod.rs
@@ -6,8 +6,6 @@ pub use reset_cause::*;
 mod serial_number;
 pub use serial_number::*;
 
-pub mod rtc_timer;
-
 #[cfg(feature = "unproven")]
 pub mod adc;
